requirements.txt
algoliasearch-django>=3.0,<4.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.15
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.7
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
typing-extensions==4.12.2
urllib3==2.2.2
